# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polymarket Trading API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MarketStatus = (typeof MarketStatus)[keyof typeof MarketStatus];

export const MarketStatus = {
  active: "active",
  resolved: "resolved",
  closed: "closed",
} as const;

export interface Market {
  id: string;
  question: string;
  category: string;
  status: MarketStatus;
  yesPrice: number;
  noPrice: number;
  volume: number;
  volume24h: number;
  liquidity: number;
  endDate: string;
  resolvedOutcome?: string | null;
  description?: string;
  conditionId?: string;
  tokenId?: string;
}

export type PositionSide = (typeof PositionSide)[keyof typeof PositionSide];

export const PositionSide = {
  YES: "YES",
  NO: "NO",
} as const;

export interface Position {
  id: string;
  marketId: string;
  marketQuestion: string;
  side: PositionSide;
  shares: number;
  avgPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPercent: number;
  value: number;
}

export type OrderSide = (typeof OrderSide)[keyof typeof OrderSide];

export const OrderSide = {
  YES: "YES",
  NO: "NO",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  open: "open",
  filled: "filled",
  cancelled: "cancelled",
  partial: "partial",
} as const;

export interface Order {
  id: string;
  marketId: string;
  marketQuestion: string;
  side: OrderSide;
  type: OrderType;
  price: number;
  amount: number;
  shares: number;
  status: OrderStatus;
  createdAt: string;
}

export type PlaceOrderRequestSide =
  (typeof PlaceOrderRequestSide)[keyof typeof PlaceOrderRequestSide];

export const PlaceOrderRequestSide = {
  YES: "YES",
  NO: "NO",
} as const;

export type PlaceOrderRequestType =
  (typeof PlaceOrderRequestType)[keyof typeof PlaceOrderRequestType];

export const PlaceOrderRequestType = {
  BUY: "BUY",
  SELL: "SELL",
} as const;

export interface PlaceOrderRequest {
  marketId: string;
  side: PlaceOrderRequestSide;
  type: PlaceOrderRequestType;
  price: number;
  amount: number;
}

export interface PortfolioSummary {
  totalValue: number;
  availableBalance: number;
  investedAmount: number;
  totalPnl: number;
  totalPnlPercent: number;
  openPositions: number;
  totalTrades: number;
  winRate: number;
}

export interface PnlPoint {
  date: string;
  pnl: number;
  cumulative: number;
}

export type OpportunityRecommendedSide =
  (typeof OpportunityRecommendedSide)[keyof typeof OpportunityRecommendedSide];

export const OpportunityRecommendedSide = {
  YES: "YES",
  NO: "NO",
} as const;

export type OpportunityRiskLevel =
  (typeof OpportunityRiskLevel)[keyof typeof OpportunityRiskLevel];

export const OpportunityRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Opportunity {
  marketId: string;
  question: string;
  category: string;
  recommendedSide: OpportunityRecommendedSide;
  currentPrice: number;
  estimatedFairValue: number;
  edge: number;
  expectedReturn: number;
  kellyFraction: number;
  suggestedAmount: number;
  riskLevel: OpportunityRiskLevel;
  daysToResolution: number;
  volume24h: number;
  rationale: string;
  conditionId?: string;
}

export interface StrategyConfig {
  autoTradingEnabled: boolean;
  bankroll: number;
  maxPositionPct: number;
  minEdge: number;
  minProbability: number;
  maxDaysToResolution: number;
  minVolume24h: number;
  scanIntervalMinutes: number;
  telegramAlertsEnabled: boolean;
  maxDailyTrades: number;
}

export interface BacktestRequest {
  daysBack: number;
  bankroll: number;
  minProbability: number;
  maxDaysToResolution: number;
  maxPositionPct: number;
}

export type BacktestTradeSide =
  (typeof BacktestTradeSide)[keyof typeof BacktestTradeSide];

export const BacktestTradeSide = {
  YES: "YES",
  NO: "NO",
} as const;

export type BacktestTradeOutcome =
  (typeof BacktestTradeOutcome)[keyof typeof BacktestTradeOutcome];

export const BacktestTradeOutcome = {
  win: "win",
  loss: "loss",
} as const;

export interface BacktestTrade {
  date: string;
  question: string;
  side: BacktestTradeSide;
  entryPrice: number;
  exitPrice: number;
  amount: number;
  pnl: number;
  pnlPct: number;
  outcome: BacktestTradeOutcome;
}

export interface BacktestResult {
  totalReturn: number;
  totalReturnPct: number;
  winRate: number;
  totalTrades: number;
  winningTrades: number;
  losingTrades: number;
  avgReturn: number;
  maxDrawdown: number;
  sharpeRatio: number;
  trades: BacktestTrade[];
  equityCurve: PnlPoint[];
}

export type WalletStatusDataSource =
  (typeof WalletStatusDataSource)[keyof typeof WalletStatusDataSource];

export const WalletStatusDataSource = {
  live: "live",
  demo: "demo",
} as const;

export interface WalletStatus {
  connected: boolean;
  address: string | null;
  usdcBalance: number;
  hasApiCredentials: boolean;
  network: string;
  dataSource: WalletStatusDataSource;
  telegramConfigured: boolean;
}

export interface TelegramTestResult {
  success: boolean;
  message: string;
}

export type ListMarketsParams = {
  category?: string;
  status?: ListMarketsStatus;
  search?: string;
};

export type ListMarketsStatus =
  (typeof ListMarketsStatus)[keyof typeof ListMarketsStatus];

export const ListMarketsStatus = {
  active: "active",
  resolved: "resolved",
  all: "all",
} as const;