# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polymarket Trading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets:
    get:
      operationId: listMarkets
      tags: [markets]
      summary: List all markets
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of markets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Market"
...
  /markets/{marketId}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get market details
      parameters:
        - name: marketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
...
components:
  schemas:
...
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, resolved, closed]
        yesPrice:
          type: number
        noPrice:
          type: number
        volume:
          type: number
        volume24h:
          type: number
        liquidity:
          type: number
        endDate:
          type: string
          format: date-time
        resolvedOutcome:
          type: string
          nullable: true
        description:
          type: string
        conditionId:
          type: string
        tokenId:
          type: string
      required:
        - id
        - question
        - category
        - status
        - yesPrice
        - noPrice
        - volume
        - volume24h
        - liquidity
        - endDate
...